Plan level sponsors

+++ database/migrations/2017_01_22_060253_create_plans_table.php
@@ -19,9 +19,9 @@ class CreatePlansTable extends Migration
             $table->text('description')->nullable();
             $table->double('min_cost', 15, 8);
             $table->double('max_cost', 15, 8);
-            $table->double('sponsor', 15, 8);
-            $table->double('paring', 15, 8);
-            $table->double('daily', 15, 8);
+            $table->decimal('sponsor');
+            $table->decimal('paring');
+            $table->decimal('daily');
             $table->integer('duration_exp');
             $table->timestamps();
         });
